+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/STSTestUtils.hs
@@ -465,8 +465,8 @@ findMismatch Alonzo (ShelleyInAlonzoUtxowPredFailure (Shelley.UtxoFailure (Utxos
 findMismatch Babbage (Babbage.UtxoFailure (AlonzoInBabbageUtxoPredFailure (UtxosFailure x@(ValidationTagMismatch _ _)))) = Just $ injectFailure x
 findMismatch
   Conway
-  ( Babbage.UtxoFailure
-      (AlonzoInBabbageUtxoPredFailure (UtxosFailure x@(Conway.ValidationTagMismatch _ _)))
+  ( Conway.UtxoFailure
+      (Conway.UtxosFailure x@(Conway.ValidationTagMismatch _ _))
     ) = Just $ injectFailure x
 findMismatch _ _ = Nothing