+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -1172,7 +1172,7 @@ ppUTXOW Allegra x = ppShelleyUtxowPredFailure x
 ppUTXOW Mary x = ppShelleyUtxowPredFailure x
 ppUTXOW Alonzo x = ppAlonzoUtxowPredFailure x
 ppUTXOW p@Babbage x = ppBabbageUtxowPredFailure p x
-ppUTXOW p@Conway x = ppBabbageUtxowPredFailure p x
+ppUTXOW p@Conway x = ppConwayUtxowPredFailure p x
 
 ppUTXOS :: Proof era -> PredicateFailure (EraRule "UTXOS" era) -> PDoc
 ppUTXOS Alonzo x = ppUtxosPredicateFailure x
@@ -1228,7 +1228,7 @@ ppUTXO Allegra x = ppAllegraUtxoPredFailure x
 ppUTXO Mary x = ppAllegraUtxoPredFailure x
 ppUTXO Alonzo x = ppAlonzoUtxoPredFailure x
 ppUTXO Babbage x = ppBabbageUtxoPredFailure x
-ppUTXO Conway x = ppBabbageUtxoPredFailure x
+ppUTXO Conway x = ppConwayUtxoPredFailure x
 
 ppLEDGERS :: Proof era -> PredicateFailure (EraRule "LEDGERS" era) -> PDoc
 ppLEDGERS p@Shelley x = unReflect ppShelleyLedgersPredFailure p x
@@ -1300,6 +1300,107 @@ ppShelleyDelegsPredFailure p (DelplFailure x) = ppDELPL p x
 
 instance Reflect era => PrettyA (ShelleyDelegsPredFailure era) where
   prettyA = ppShelleyDelegsPredFailure reify
+
+ppConwayUtxoPredFailure ::
+  forall era.
+  Reflect era =>
+  ConwayRules.ConwayUtxoPredFailure era ->
+  PDoc
+ppConwayUtxoPredFailure = \case
+  ConwayRules.UtxosFailure yy -> ppSexp "UtxosFailure" [ppUTXOS @era reify yy]
+  ConwayRules.BadInputsUTxO txins -> ppSexp "BadInputsUTxO" [ppSet pcTxIn txins]
+  ConwayRules.OutsideValidityIntervalUTxO vi slot ->
+    ppRecord
+      "OutsideValidityIntervalUTxO"
+      [ ("provided interval", ppValidityInterval vi)
+      , ("current slot", pcSlotNo slot)
+      ]
+  ConwayRules.MaxTxSizeUTxO actual maxs ->
+    ppRecord "MaxTxSizeUTxO" [("Actual", ppInteger actual), ("max transaction size", ppInteger maxs)]
+  ConwayRules.InputSetEmptyUTxO -> ppString "InputSetEmptyUTxO"
+  ConwayRules.FeeTooSmallUTxO computed supplied ->
+    ppRecord
+      "FeeTooSmallUTxO"
+      [ ("min fee for this transaction", pcCoin computed)
+      , ("fee supplied by this transaction", pcCoin supplied)
+      ]
+  ConwayRules.ValueNotConservedUTxO consumed produced ->
+    ppRecord
+      "ValueNotConservedUTxO"
+      [("coin consumed", pcVal @era reify consumed), ("coin produced", pcVal @era reify produced)]
+  ConwayRules.WrongNetwork n add ->
+    ppRecord
+      "WrongNetwork"
+      [ ("expected network id", ppNetwork n)
+      , ("set of addresses with wrong network id", ppSet pcAddr add)
+      ]
+  ConwayRules.WrongNetworkWithdrawal n accnt ->
+    ppRecord
+      "WrongNetworkWithdrawal"
+      [ ("expected network id", ppNetwork n)
+      , ("set reward address with wrong network id", ppSet pcRewardAccount accnt)
+      ]
+  ConwayRules.OutputTooSmallUTxO xs ->
+    ppRecord
+      "OutputTooSmallUTxO"
+      [("list of supplied transaction outputs that are too small", ppList (pcTxOut reify) xs)]
+  ConwayRules.OutputBootAddrAttrsTooBig xs ->
+    ppRecord
+      "OutputBootAddrAttrsTooBig"
+      [("list of supplied bad transaction outputs", ppList (pcTxOut reify) xs)]
+  ConwayRules.OutputTooBigUTxO xs ->
+    ppSexp
+      "OutputTooBigUTxO"
+      [ ppList
+          ( \(a, b, c) ->
+              ppRecord'
+                ""
+                [("actual size", ppInt a), ("PParam max value", ppInt b), ("TxOut", pcTxOut reify c)]
+          )
+          xs
+      ]
+  ConwayRules.InsufficientCollateral c1 c2 ->
+    ppRecord
+      "InsufficientCollateral"
+      [ ("balance computed", prettyA c1)
+      , ("the required collateral for the given fee", pcCoin c2)
+      ]
+  ConwayRules.ScriptsNotPaidUTxO u -> ppSexp "ScriptsNotPaidUTxO" [pcUTxO reify u]
+  ConwayRules.ExUnitsTooBigUTxO e1 e2 ->
+    ppRecord
+      "ExUnitsTooBigUTxO"
+      [ ("Max EXUnits from the protocol parameters", pcExUnits e1)
+      , ("EXUnits supplied", pcExUnits e2)
+      ]
+  ConwayRules.CollateralContainsNonADA v -> ppSexp "CollateralContainsNonADA" [pcVal (reify @era) v]
+  ConwayRules.WrongNetworkInTxBody n1 n2 ->
+    ppRecord
+      "WrongNetworkInTxBody"
+      [ ("Actual Network ID", ppNetwork n1)
+      , ("Network ID in transaction body", ppNetwork n2)
+      ]
+  ConwayRules.OutsideForecast slot -> ppRecord "OutsideForecast" [("slot number outside consensus forecast range", pcSlotNo slot)]
+  ConwayRules.TooManyCollateralInputs n1 n2 ->
+    ppRecord
+      "TooManyCollateralInputs"
+      [ ("Max allowed collateral inputs", ppNatural n1)
+      , ("Number of collateral inputs", ppNatural n2)
+      ]
+  ConwayRules.NoCollateralInputs -> ppSexp " NoCollateralInputs" []
+  ConwayRules.IncorrectTotalCollateralField c1 c2 ->
+    ppRecord
+      "IncorrectTotalCollateralField"
+      [("collateral provided", prettyA c1), ("collateral declared", pcCoin c2)]
+  ConwayRules.BabbageOutputTooSmallUTxO xs ->
+    ppSexp "BabbageOutputTooSmallUTxO" [ppList (ppPair (pcTxOut reify) pcCoin) xs]
+  ConwayRules.BabbageNonDisjointRefInputs xs ->
+    ppSexp "BabbageNonDisjointRefInputs" [ppList pcTxIn (toList xs)]
+
+instance
+  Reflect era =>
+  PrettyA (ConwayRules.ConwayUtxoPredFailure era)
+  where
+  prettyA = ppConwayUtxoPredFailure
 
 ppBabbageUtxoPredFailure :: Reflect era => BabbageUtxoPredFailure era -> PDoc
 ppBabbageUtxoPredFailure (AlonzoInBabbageUtxoPredFailure x) = ppAlonzoUtxoPredFailure x
@@ -1416,6 +1517,66 @@ ppShelleyLedgerPredFailure proof (DelegsFailure x) = ppDELEGS proof x
 
 instance Reflect era => PrettyA (ShelleyLedgerPredFailure era) where
   prettyA = ppShelleyLedgerPredFailure reify
+
+ppConwayUtxowPredFailure ::
+  Reflect era =>
+  Proof era ->
+  ConwayRules.ConwayUtxowPredFailure era ->
+  PDoc
+ppConwayUtxowPredFailure proof = \case
+  ConwayRules.UtxoFailure x -> ppSexp "UtxoFailure" [ppUTXO proof x]
+  ConwayRules.InvalidWitnessesUTXOW vkeyws ->
+    ppSexp "InvalidWitnessesUTXOW" [ppList ppVKey vkeyws]
+  ConwayRules.MissingVKeyWitnessesUTXOW whs ->
+    ppSexp "MissingVKeyWitnessesUTXOW" [ppWitHashes whs]
+  ConwayRules.MissingScriptWitnessesUTXOW m ->
+    ppSexp "MissingScriptWitnessesUTXOW" [ppSet pcScriptHash m]
+  ConwayRules.ScriptWitnessNotValidatingUTXOW m ->
+    ppSexp "ScriptWitnessNotValidatingUTXOW" [ppSet pcScriptHash m]
+  ConwayRules.MissingTxBodyMetadataHash m ->
+    ppSexp " MissingTxMetadata" [ppAuxiliaryDataHash m]
+  ConwayRules.MissingTxMetadata m ->
+    ppSexp " MissingTxMetadata" [ppAuxiliaryDataHash m]
+  ConwayRules.ConflictingMetadataHash h1 h2 ->
+    ppRecord
+      "ConflictingMetadataHash"
+      [("Hash in the body", ppAuxiliaryDataHash h1), ("Hash of full metadata", ppAuxiliaryDataHash h2)]
+  ConwayRules.InvalidMetadata ->
+    ppSexp "InvalidMetadata" []
+  ConwayRules.ExtraneousScriptWitnessesUTXOW m ->
+    ppSexp "ExtraneousScriptWitnessesUTXOW" [ppSet pcScriptHash m]
+  ConwayRules.MissingRedeemers xs ->
+    ppSexp "MissingRedeemers" [ppList (ppPair ppPlutusPurposeAsItem prettyA) xs]
+  ConwayRules.MissingRequiredDatums s1 s2 ->
+    ppRecord
+      "MissingRequiredDatums"
+      [ ("missing data hashes", ppSet ppSafeHash s1)
+      , ("received data hashes", ppSet ppSafeHash s2)
+      ]
+  ConwayRules.NotAllowedSupplementalDatums s1 s2 ->
+    ppRecord
+      "NotAllowedSupplementalDatums"
+      [ ("unallowed data hashes", ppSet ppSafeHash s1)
+      , ("acceptable data hashes", ppSet ppSafeHash s2)
+      ]
+  ConwayRules.PPViewHashesDontMatch h1 h2 ->
+    ppRecord
+      "PPViewHashesDontMatch"
+      [ ("PPHash in the TxBody", ppStrictMaybe ppSafeHash h1)
+      , ("PPHash Computed from the current Protocol Parameters", ppStrictMaybe ppSafeHash h2)
+      ]
+  ConwayRules.UnspendableUTxONoDatumHash x ->
+    ppSexp "UnspendableUTxONoDatumHash" [ppSet pcTxIn x]
+  ConwayRules.ExtraRedeemers x ->
+    ppSexp "ExtraRedeemers" [ppList ppPlutusPurposeAsIx x]
+  ConwayRules.MalformedScriptWitnesses x -> ppSexp "MalformedScriptWitnesses" [ppSet pcScriptHash x]
+  ConwayRules.MalformedReferenceScripts x -> ppSexp "MalformedReferenceScripts" [ppSet pcScriptHash x] -- !(Set (ScriptHash (EraCrypto era)))
+
+instance
+  Reflect era =>
+  PrettyA (ConwayRules.ConwayUtxowPredFailure era)
+  where
+  prettyA = ppConwayUtxowPredFailure reify
 
 ppBabbageUtxowPredFailure ::
   Reflect era =>